Fix polling namespace on compute nodes
Previously only ceilometer-polling services running in the central
container had a polling_namespace configured. Those running on compute
nodes used the default which is to run all pollsters from the central
and compute namespace. But there is no need to gather the data for the
central namespace on every compute node.

With this fix the central container only runs the central pollsters and
those on the compute nodes only the compute pollsters. This can be
adjusted by config_template overrides in group_vars if needed.

Change-Id: Ia130bd7f9000882b556aea2190de97ddb290bafc