Update role for new source build process
The variables ceilometer_developer_mode and ceilometer_venv_download
no longer carry any meaning. This review changes ceilometer to
do the equivalent of what developer_mode was all the time,
meaning that it always builds the venv and never requires
the repo server, but it will use a repo server when available.

As part of this, we move the source build out of its own file
because it's now a single task to include the venv build role.
This is just to make it easier to follow the code.

We also change include_tasks to import_tasks and include_role
to import_role so that the tags in the python_venv_build role
will work.

Change-Id: Ib6423249ad6d8c24382a0d82229f766d34255ee9